requirements.txt
pygame-ce >= 2.3.0
tomlkit >= 0.12.3
